celery_log_docker/pyproject.toml
[tool.poetry]
name = "celery_log"
version = "0.1.0"
description = "testing log for celery in file"
authors = ["Dmitry Afanasyev <Balshbox@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.9"
celery = "^5.1.2"
ipython = "^7.25"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
